flask-server-test/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
ortools==9.12.4544
python-dotenv==1.0.1
